LayoutTests/animations/animation-shorthand-expected.txt

Testing webkitAnimationName on a: PASS
Testing webkitAnimationDuration on a: PASS
Testing webkitAnimationTimingFunction on a: PASS
Testing webkitAnimationDelay on a: PASS
Testing webkitAnimationIterationCount on a: PASS
Testing webkitAnimationDirection on a: PASS
Testing webkitAnimationFillMode on a: PASS
Testing webkitAnimationName on b: PASS
Testing webkitAnimationDuration on b: PASS
Testing webkitAnimationTimingFunction on b: PASS
Testing webkitAnimationDelay on b: PASS
Testing webkitAnimationIterationCount on b: PASS
Testing webkitAnimationDirection on b: PASS
Testing webkitAnimationFillMode on b: PASS
Testing webkitAnimationName on c: PASS
Testing webkitAnimationDuration on c: PASS
Testing webkitAnimationTimingFunction on c: PASS
Testing webkitAnimationDelay on c: PASS
Testing webkitAnimationIterationCount on c: PASS
Testing webkitAnimationDirection on c: PASS
Testing webkitAnimationFillMode on c: PASS
Testing webkitAnimationName on d: PASS
Testing webkitAnimationDuration on d: PASS
Testing webkitAnimationTimingFunction on d: PASS
Testing webkitAnimationDelay on d: PASS
Testing webkitAnimationIterationCount on d: PASS
Testing webkitAnimationDirection on d: PASS
Testing webkitAnimationFillMode on d: PASS
Testing webkitAnimationName on e: PASS
Testing webkitAnimationDuration on e: PASS
Testing webkitAnimationTimingFunction on e: PASS
Testing webkitAnimationDelay on e: PASS
Testing webkitAnimationIterationCount on e: PASS
Testing webkitAnimationDirection on e: PASS
Testing webkitAnimationFillMode on e: PASS
Testing webkitAnimationName on f: PASS
Testing webkitAnimationDuration on f: PASS
Testing webkitAnimationTimingFunction on f: PASS
Testing webkitAnimationDelay on f: PASS
Testing webkitAnimationIterationCount on f: PASS
Testing webkitAnimationDirection on f: PASS
Testing webkitAnimationFillMode on f: PASS
Testing webkitAnimationName on g: PASS
Testing webkitAnimationDuration on g: PASS
Testing webkitAnimationTimingFunction on g: PASS
Testing webkitAnimationDelay on g: PASS
Testing webkitAnimationIterationCount on g: PASS
Testing webkitAnimationDirection on g: PASS
Testing webkitAnimationFillMode on g: PASS
Testing webkitAnimationName on h: PASS
Testing webkitAnimationDuration on h: PASS
Testing webkitAnimationTimingFunction on h: PASS
Testing webkitAnimationDelay on h: PASS
Testing webkitAnimationIterationCount on h: PASS
Testing webkitAnimationDirection on h: PASS
Testing webkitAnimationFillMode on h: PASS
Testing webkitAnimationName on i: PASS
Testing webkitAnimationDuration on i: PASS
Testing webkitAnimationTimingFunction on i: PASS
Testing webkitAnimationDelay on i: PASS
Testing webkitAnimationIterationCount on i: PASS
Testing webkitAnimationDirection on i: PASS
Testing webkitAnimationFillMode on i: PASS
Testing webkitAnimationName on j: PASS
Testing webkitAnimationDuration on j: PASS
Testing webkitAnimationTimingFunction on j: PASS
Testing webkitAnimationDelay on j: PASS
Testing webkitAnimationIterationCount on j: PASS
Testing webkitAnimationDirection on j: PASS
Testing webkitAnimationFillMode on j: PASS
